+++ src/util.js
@@ -23,9 +23,9 @@ define(function (require) {
      *
      * 传入一个字符串时，只返回一个发送器函数；传入数组或对象时，递归；传入函数时
      *
-     * @param {string|Array.<string>|Object.<string, string>|function} url 请求路径或多个请求路径的集合，或是取值函数
+     * @param {string|Array.<string>|Object.<string, string>|Function} url 请求路径或多个请求路径的集合，或是取值函数
      * @param {function(string):boolean} isRequester 判断是否是需要生成请求发送器的路径
-     * @return {function|Array.<function>|Object.<string, function>} 将对应的路径转换为发送器后返回
+     * @return {Function|Array.<Function>|Object.<string, Function>} 将对应的路径转换为发送器后返回
      */
     util.genRequesters = function (url, isRequester) {
         if (u.typeOf(url) === 'String') {