+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 * 0.2.6
+    - 增加 `system/session` 模块
+    - `ListView` 原型增加 `batchModify` 方法，提供批量操作在 View 层的扩展点
+    - 增加表单元素的字体默认设置
+    - `tpl` 模块支持解析 `data-ui` 属性中以集中方式设置的 `type`
+    - 修复 `.list-operation` 样式在 Firefox 下的问题
     - 修正代码规范
 
 * 0.2.5