+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - 增加表单元素的字体默认设置
     - `tpl` 模块支持解析 `data-ui` 属性中以集中方式设置的 `type`
     - 修复 `.list-operation` 样式在 Firefox 下的问题
+    - 调整了图表控件中 `echarts` 的 `formatter` 参数
     - 修正代码规范
 
 * 0.2.5